add setuptools packaging
