+++ setup.cfg
@@ -1,0 +1,49 @@
+[metadata]
+name = SimConnect
+version = attr: PythonSimConnect.__version__
+description = Adds a pythonic wrapper for SimConnect SDK.
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/odwdinc/Python-SimConnect
+author = Peter Heiss
+author_email = peter.heiss@uni-muenster.de
+maintainer = Peter Heiss
+maintainer_email = peter.heiss@uni-muenster.de
+keywords = Ctypes, FlightSim, SimConnect
+license = MIT
+classifiers =
+    Development Status :: 2 - Pre-Alpha
+    Environment :: Console
+    Intended Audience :: Developers
+    License :: OSI Approved :: MIT License
+    Operating System :: OS Independent
+    Programming Language :: Python
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3.6
+    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: Implementation :: CPython
+
+
+[options]
+packages = SimConnect
+python_requires = >=3.6
+include_package_data=True
+test_suite="tests"
+
+[flake8]
+exclude = .venv,.tox,dist,docs,build,*.egg,env,venv,.undodir
+
+[bdist_wheel]
+universal = 1
+
+[build-system]
+requires = ["setuptools >= 40.6.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[tool:pytest]
+minversion = 5.0
+addopts = -ra -q
+testpaths =
+    tests
+log_level = DEBUG
